Extract package downloading code
