+++ src/makePatch.ts
@@ -1,34 +1,31 @@
 import chalk from "chalk"
-import { join, dirname, resolve } from "./path"
-import { spawnSafeSync } from "./spawnSafe"
-import { PackageManager } from "./detectPackageManager"
-import { removeIgnoredFiles } from "./filterFiles"
 import {
-  writeFileSync,
+  copySync,
   existsSync,
   mkdirSync,
-  unlinkSync,
-  mkdirpSync,
   realpathSync,
+  unlinkSync,
+  writeFileSync,
 } from "fs-extra"
 import { sync as rimraf } from "rimraf"
-import { copySync } from "fs-extra"
-import { dirSync } from "tmp"
-import { getPatchFiles } from "./patchFs"
-import {
-  getPatchDetailsFromCliString,
-  getPackageDetailsFromPatchFilename,
-  PackageDetails,
-} from "./PackageDetails"
-import { resolveRelativeFileDependencies } from "./resolveRelativeFileDependencies"
-import { getPackageResolution } from "./getPackageResolution"
-import { parsePatchFile } from "./patch/parse"
 import { gzipSync } from "zlib"
-import { getPackageVersion } from "./getPackageVersion"
 import {
   maybePrintIssueCreationPrompt,
   openIssueCreationLink,
 } from "./createIssue"
+import { PackageManager } from "./detectPackageManager"
+import { downloadPackage } from "./downloadPackage"
+import { removeIgnoredFiles } from "./filterFiles"
+import { getPackageVersion } from "./getPackageVersion"
+import {
+  getPackageDetailsFromPatchFilename,
+  getPatchDetailsFromCliString,
+  PackageDetails,
+} from "./PackageDetails"
+import { parsePatchFile } from "./patch/parse"
+import { getPatchFiles } from "./patchFs"
+import { dirname, join, resolve } from "./path"
+import { spawnSafeSync } from "./spawnSafe"
 
 function printNoPackageFoundError(
   packageName: string,
@@ -64,7 +61,7 @@ export function makePatch({
     console.error("No such package", packagePathSpecifier)
     return
   }
-  const appPackageJson = require(join(appPath, "package.json"))
+
   const packagePath = join(appPath, packageDetails.path)
   const packageJsonPath = join(packagePath, "package.json")
 
@@ -73,187 +70,104 @@ export function makePatch({
     process.exit(1)
   }
 
-  const tmpRepo = dirSync({ unsafeCleanup: true })
-  const tmpRepoPackagePath = join(tmpRepo.name, packageDetails.path)
-  const tmpRepoNpmRoot = tmpRepoPackagePath.slice(
-    0,
-    -`/node_modules/${packageDetails.name}`.length,
+  const { tmpRepo, tmpRepoPackagePath } = downloadPackage({
+    packageDetails,
+    appPath,
+    packageManager,
+  })
+
+  const patchesDir = resolve(join(appPath, patchDir))
+  const packageVersion = getPackageVersion(
+    join(resolve(packageDetails.path), "package.json"),
   )
 
-  const tmpRepoPackageJsonPath = join(tmpRepoNpmRoot, "package.json")
-
-  try {
-    const patchesDir = resolve(join(appPath, patchDir))
-
-    console.info(chalk.grey("•"), "Creating temporary folder")
-
-    // make a blank package.json
-    mkdirpSync(tmpRepoNpmRoot)
-    writeFileSync(
-      tmpRepoPackageJsonPath,
-      JSON.stringify({
-        dependencies: {
-          [packageDetails.name]: getPackageResolution({
-            packageDetails,
-            packageManager,
-            appPath,
-          }),
-        },
-        resolutions: resolveRelativeFileDependencies(
-          appPath,
-          appPackageJson.resolutions || {},
-        ),
-      }),
-    )
-
-    const packageVersion = getPackageVersion(
-      join(resolve(packageDetails.path), "package.json"),
-    )
-
-    // copy .npmrc/.yarnrc in case packages are hosted in private registry
-    // tslint:disable-next-line:align
-    ;[".npmrc", ".yarnrc"].forEach((rcFile) => {
-      const rcPath = join(appPath, rcFile)
-      if (existsSync(rcPath)) {
-        copySync(rcPath, join(tmpRepo.name, rcFile))
-      }
+  const git = (...args: string[]) =>
+    spawnSafeSync("git", args, {
+      cwd: tmpRepo.name,
+      env: { ...process.env, HOME: tmpRepo.name },
+      maxBuffer: 1024 * 1024 * 100,
     })
 
-    if (packageManager === "yarn") {
-      console.info(
-        chalk.grey("•"),
-        `Installing ${packageDetails.name}@${packageVersion} with yarn`,
-      )
-      try {
-        // try first without ignoring scripts in case they are required
-        // this works in 99.99% of cases
-        spawnSafeSync(`yarn`, ["install", "--ignore-engines"], {
-          cwd: tmpRepoNpmRoot,
-          logStdErrOnError: false,
-        })
-      } catch (e) {
-        // try again while ignoring scripts in case the script depends on
-        // an implicit context which we havn't reproduced
-        spawnSafeSync(
-          `yarn`,
-          ["install", "--ignore-engines", "--ignore-scripts"],
-          {
-            cwd: tmpRepoNpmRoot,
-          },
-        )
-      }
-    } else {
-      console.info(
-        chalk.grey("•"),
-        `Installing ${packageDetails.name}@${packageVersion} with npm`,
-      )
-      try {
-        // try first without ignoring scripts in case they are required
-        // this works in 99.99% of cases
-        spawnSafeSync(`npm`, ["i", "--force"], {
-          cwd: tmpRepoNpmRoot,
-          logStdErrOnError: false,
-          stdio: "ignore",
-        })
-      } catch (e) {
-        // try again while ignoring scripts in case the script depends on
-        // an implicit context which we havn't reproduced
-        spawnSafeSync(`npm`, ["i", "--ignore-scripts", "--force"], {
-          cwd: tmpRepoNpmRoot,
-          stdio: "ignore",
-        })
-      }
-    }
+  // remove nested node_modules just to be safe
+  rimraf(join(tmpRepoPackagePath, "node_modules"))
+  // remove .git just to be safe
+  rimraf(join(tmpRepoPackagePath, ".git"))
 
-    const git = (...args: string[]) =>
-      spawnSafeSync("git", args, {
-        cwd: tmpRepo.name,
-        env: { ...process.env, HOME: tmpRepo.name },
-        maxBuffer: 1024 * 1024 * 100,
-      })
+  // commit the package
+  console.info(chalk.grey("•"), "Diffing your files with clean files")
+  writeFileSync(join(tmpRepo.name, ".gitignore"), "!/node_modules\n\n")
+  git("init")
+  git("config", "--local", "user.name", "patch-package")
+  git("config", "--local", "user.email", "patch@pack.age")
 
-    // remove nested node_modules just to be safe
-    rimraf(join(tmpRepoPackagePath, "node_modules"))
-    // remove .git just to be safe
-    rimraf(join(tmpRepoPackagePath, ".git"))
+  // remove ignored files first
+  removeIgnoredFiles(tmpRepoPackagePath, includePaths, excludePaths)
 
-    // commit the package
-    console.info(chalk.grey("•"), "Diffing your files with clean files")
-    writeFileSync(join(tmpRepo.name, ".gitignore"), "!/node_modules\n\n")
-    git("init")
-    git("config", "--local", "user.name", "patch-package")
-    git("config", "--local", "user.email", "patch@pack.age")
+  git("add", "-f", packageDetails.path)
+  git("commit", "--allow-empty", "-m", "init")
 
-    // remove ignored files first
-    removeIgnoredFiles(tmpRepoPackagePath, includePaths, excludePaths)
+  // replace package with user's version
+  rimraf(tmpRepoPackagePath)
 
-    git("add", "-f", packageDetails.path)
-    git("commit", "--allow-empty", "-m", "init")
+  // pnpm installs packages as symlinks, copySync would copy only the symlink
+  copySync(realpathSync(packagePath), tmpRepoPackagePath)
 
-    // replace package with user's version
-    rimraf(tmpRepoPackagePath)
+  // remove nested node_modules just to be safe
+  rimraf(join(tmpRepoPackagePath, "node_modules"))
+  // remove .git just to be safe
+  rimraf(join(tmpRepoPackagePath, ".git"))
 
-    // pnpm installs packages as symlinks, copySync would copy only the symlink
-    copySync(realpathSync(packagePath), tmpRepoPackagePath)
+  // also remove ignored files like before
+  removeIgnoredFiles(tmpRepoPackagePath, includePaths, excludePaths)
 
-    // remove nested node_modules just to be safe
-    rimraf(join(tmpRepoPackagePath, "node_modules"))
-    // remove .git just to be safe
-    rimraf(join(tmpRepoPackagePath, ".git"))
+  // stage all files
+  git("add", "-f", packageDetails.path)
 
-    // also remove ignored files like before
-    removeIgnoredFiles(tmpRepoPackagePath, includePaths, excludePaths)
+  // get diff of changes
+  const diffResult = git(
+    "diff",
+    "--cached",
+    "--no-color",
+    "--ignore-space-at-eol",
+    "--no-ext-diff",
+  )
 
-    // stage all files
-    git("add", "-f", packageDetails.path)
-
-    // get diff of changes
-    const diffResult = git(
-      "diff",
-      "--cached",
-      "--no-color",
-      "--ignore-space-at-eol",
-      "--no-ext-diff",
+  if (diffResult.stdout.length === 0) {
+    console.warn(
+      `⁉️  Not creating patch file for package '${packagePathSpecifier}'`,
     )
+    console.warn(`⁉️  There don't appear to be any changes.`)
+    process.exit(1)
+    return
+  }
 
-    if (diffResult.stdout.length === 0) {
-      console.warn(
-        `⁉️  Not creating patch file for package '${packagePathSpecifier}'`,
-      )
-      console.warn(`⁉️  There don't appear to be any changes.`)
-      process.exit(1)
-      return
-    }
-
-    try {
-      parsePatchFile(diffResult.stdout.toString())
-    } catch (e) {
-      if (
-        (e as Error).message.includes("Unexpected file mode string: 120000")
-      ) {
-        console.error(`
+  try {
+    parsePatchFile(diffResult.stdout.toString())
+  } catch (e) {
+    if ((e as Error).message.includes("Unexpected file mode string: 120000")) {
+      console.error(`
 ⛔️ ${chalk.red.bold("ERROR")}
 
   Your changes involve creating symlinks. patch-package does not yet support
   symlinks.
   
   ️Please use ${chalk.bold("--include")} and/or ${chalk.bold(
-          "--exclude",
-        )} to narrow the scope of your patch if
+        "--exclude",
+      )} to narrow the scope of your patch if
   this was unintentional.
 `)
-      } else {
-        const outPath = "./patch-package-error.json.gz"
-        writeFileSync(
-          outPath,
-          gzipSync(
-            JSON.stringify({
-              error: { message: e.message, stack: e.stack },
-              patch: diffResult.stdout.toString(),
-            }),
-          ),
-        )
-        console.error(`
+    } else {
+      const outPath = "./patch-package-error.json.gz"
+      writeFileSync(
+        outPath,
+        gzipSync(
+          JSON.stringify({
+            error: { message: e.message, stack: e.stack },
+            patch: diffResult.stdout.toString(),
+          }),
+        ),
+      )
+      console.error(`
 ⛔️ ${chalk.red.bold("ERROR")}
         
   patch-package was unable to read the patch-file made by git. This should not
@@ -271,47 +185,41 @@ export function makePatch({
   attempting to patch.
 
 `)
-      }
-      process.exit(1)
-      return
     }
+    process.exit(1)
+    return
+  }
 
-    // maybe delete existing
-    getPatchFiles(patchDir).forEach((filename) => {
-      const deets = getPackageDetailsFromPatchFilename(filename)
-      if (deets && deets.path === packageDetails.path) {
-        unlinkSync(join(patchDir, filename))
-      }
-    })
+  // maybe delete existing
+  getPatchFiles(patchDir).forEach((filename) => {
+    const deets = getPackageDetailsFromPatchFilename(filename)
+    if (deets && deets.path === packageDetails.path) {
+      unlinkSync(join(patchDir, filename))
+    }
+  })
 
-    const patchFileName = createPatchFileName({
+  const patchFileName = createPatchFileName({
+    packageDetails,
+    packageVersion,
+  })
+
+  const patchPath = join(patchesDir, patchFileName)
+  if (!existsSync(dirname(patchPath))) {
+    // scoped package
+    mkdirSync(dirname(patchPath))
+  }
+  writeFileSync(patchPath, diffResult.stdout)
+  console.log(
+    `${chalk.green("✔")} Created file ${join(patchDir, patchFileName)}\n`,
+  )
+  if (createIssue) {
+    openIssueCreationLink({
       packageDetails,
+      patchFileContents: diffResult.stdout.toString(),
       packageVersion,
     })
-
-    const patchPath = join(patchesDir, patchFileName)
-    if (!existsSync(dirname(patchPath))) {
-      // scoped package
-      mkdirSync(dirname(patchPath))
-    }
-    writeFileSync(patchPath, diffResult.stdout)
-    console.log(
-      `${chalk.green("✔")} Created file ${join(patchDir, patchFileName)}\n`,
-    )
-    if (createIssue) {
-      openIssueCreationLink({
-        packageDetails,
-        patchFileContents: diffResult.stdout.toString(),
-        packageVersion,
-      })
-    } else {
-      maybePrintIssueCreationPrompt(packageDetails, packageManager)
-    }
-  } catch (e) {
-    console.error(e)
-    throw e
-  } finally {
-    tmpRepo.removeCallback()
+  } else {
+    maybePrintIssueCreationPrompt(packageDetails, packageManager)
   }
 }
